requirements.txt
asgiref==3.8.1
black==24.10.0
click==8.1.7
Django==5.1.2
djangorestframework==3.15.2
drf-yasg==1.21.8
inflection==0.5.1
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
psycopg2==2.9.10
pytz==2024.2
PyYAML==6.0.2
setuptools==75.3.0
sqlparse==0.5.1
tzdata==2024.2
uritemplate==4.1.1
wheel==0.44.0
